move from const generic to iterator approach
This implies that we parse the phdr every time when we iterator over
them however it also gives the flexibility of not providing an upper
bound of supported load segments during compile time.
Trading flexibility vs potential repetitive work is totally fine as in
use cases this crate is used, the phdrs are usually iterated once or
twice.